Phase 4: decompose tasks ResourcesTable into rows, hook, and pure helpers
