+++ packages/client/src/components/tasks/resourceMeta.ts
@@ -1,4 +1,9 @@
-import type { TaskResourceLevel } from "@emstack/types";
+import type {
+  Module,
+  ModuleGroup,
+  TaskResource,
+  TaskResourceLevel,
+} from "@emstack/types";
 
 export const RESOURCE_LEVEL_OPTIONS: { value: TaskResourceLevel;
   label: string; }[] = [
@@ -38,4 +43,49 @@ export function getResourceLevelClass(
     default:
       return "bg-muted text-muted-foreground border-muted-foreground/30";
   }
+}
+
+export type ResourceLevelKey
+  = "easeOfStarting" | "timeNeeded" | "interactivity";
+
+// The ease/time/interactivity values live on Resource / ModuleGroup / Module.
+// Resolve from most-specific to least-specific:
+//   Module → its parent ModuleGroup → Resource.
+// When the row narrows to a Module that doesn't carry an override, walk up
+// to the module's parent group (looked up via allModules/allModuleGroups
+// since the row's joined moduleGroup may be null when only a module was
+// selected) and then to the resource.
+export function inheritedLevel(
+  resource: TaskResource,
+  key: ResourceLevelKey,
+  allModules: Module[],
+  allModuleGroups: ModuleGroup[],
+): TaskResourceLevel | null {
+  const fromModule = resource.module?.[key];
+  if (fromModule) return fromModule;
+
+  if (resource.moduleId) {
+    const fullModule = allModules.find(m => m.id === resource.moduleId);
+    if (fullModule?.moduleGroupId) {
+      const parentGroup = allModuleGroups.find(
+        g => g.id === fullModule.moduleGroupId,
+      );
+      const fromParent = parentGroup?.[key];
+      if (fromParent) return fromParent;
+    }
+  }
+
+  const fromRowGroup = resource.moduleGroup?.[key];
+  if (fromRowGroup) return fromRowGroup;
+
+  return resource.resource?.[key] ?? null;
+}
+
+// Most-specific first: Module > Module Group > Resource.
+export function linkedResourceLabel(resource: TaskResource): string | null {
+  return resource.resource
+    ? [resource.module?.name, resource.moduleGroup?.name, resource.resource.name]
+      .filter(Boolean)
+      .join(" > ")
+    : null;
 }
